requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudinary==1.31.0
cryptography==41.0.2
Django==4.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-taggit==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.5
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.14
